+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 * [#5994](https://github.com/xmake-io/xmake/issues/5994): Profile process performance
 * [#5995](https://github.com/xmake-io/xmake/pull/5995): Add profile support for vs generator
 * [#5949](https://github.com/xmake-io/xmake/pull/5949): Add rule nodejs.module
+* [#3380](https://github.com/xmake-io/xmake/issues/3380): Add namespace support
+* [#5945](https://github.com/xmake-io/xmake/issues/5945): check pkgconfig/cmake importfiles for package
+
+### Bugs fixed
+
+* [#272](https://github.com/tboox/tbox/issues/272): Fix xmake stuck when readfile always return zero for msvc + /O1
 
 ## v2.9.7
 
@@ -1931,6 +1937,12 @@
 * [#5994](https://github.com/xmake-io/xmake/issues/5994): 分析进程执行性能
 * [#5995](https://github.com/xmake-io/xmake/pull/5995): 为 vs generator 添加 profile 支持
 * [#5949](https://github.com/xmake-io/xmake/pull/5949): 添加 nodejs.module 规则
+* [#3380](https://github.com/xmake-io/xmake/issues/3380): 添加命名空间支持
+* [#5945](https://github.com/xmake-io/xmake/issues/5945): 检测 pkgconfig/cmake 导入文件
+
+### Bugs 修复
+
+* [#272](https://github.com/tboox/tbox/issues/272): 修复 msvc + /O1 时候，错误的编译器优化导致 xmake 加载卡住
 
 ## v2.9.7